+++ tests/integration/test_integration_inter_session_mixed_modality.py
@@ -69,7 +69,7 @@ def test_mixed_modality(tmp_path):
         base_dir=base_dir,
         selected_folders=[npm_folder],
         store_id_to_store_label=npm_store_id_to_store_label,
-        npm_split_events=[True, True],
+        npm_split_events=[False, True],
     )
     step1(
         base_dir=base_dir,
@@ -194,7 +194,7 @@ def test_mixed_modality_tdt_npm(tmp_path):
             "file0_chod1": "signal_region1",
             "eventTrue": "ttl_true_region1",
         },
-        npm_split_events=[True, True],
+        npm_split_events=[False, True],
     )
 
     selected_folders = [str(tdt_session), str(npm_session)]
@@ -424,7 +424,7 @@ def test_mixed_modality_nwb_npm(tmp_path):
             "file0_chod1": "signal_region1",
             "eventTrue": "ttl_true_region1",
         },
-        npm_split_events=[True, True],
+        npm_split_events=[False, True],
     )
 
     selected_folders = [str(nwb_session), str(npm_session)]